requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
openai
sqlalchemy
psycopg2-binary
streamlit
pandas
plotly
sqlparse
sqlglot
chromadb
pyyaml
python-dotenv
